Move problem logic to problem_manager module
Code pertaining to handling problems was moved from the main app's
view.py to the problem_manager module. This cleans up the organization
a little bit. The original contents problem_manager were deleted (were
not being used). Also basic progress on uploading new problems was made.
